requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
cryptography==46.0.5
distro==1.9.0
docopt==0.6.2
Flask==3.1.3
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
Js2Py==0.74
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
MouseInfo==0.1.3
multidict==6.7.1
narwhals==2.15.0
numpy==2.4.1
openai==2.21.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pipwin==0.5.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.53.0
streamlit-geolocation==0.0.10
streamlit-js-eval==1.0.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.6
wikipedia==1.4.0
yarl==1.22.0
